Fix parallel build
Parallel build was failing while building the man pages. doc/Makefile.am
tries cleverly to handle the fact that c2man produces all the man pages
at once, following the approach described in:
https://www.gnu.org/software/automake/manual/html_node/Multiple-Outputs.html

However this is failing because c2man is not generating all the
requested man pages, so the rule ends being run multiple times instead
of once and parallel builds are broken. One of the requested pages is
for a function that doesn’t exist, fribidi_log2vis_get_embedding_levels_ex,
so I removed it from fribidi.def.

Two functions were not documented, so I added minimal documentation for
them.

But the last two are for variables not functions and c2man does not like
that it seems so I added a hack to drop these two as well without
modifying fribidi.def.

Fixes https://github.com/fribidi/fribidi/issues/45