Add support for building with the Meson build system
The Meson build system is being adopted by projects
such as GNOME, GTK, GStreamer, and there is support
for it in Pango as well.

Having support for building with Meson in fribidi
would facilitate builds of GTK, Pango, etc. via
Meson's subproject support.

Meson also supports builds on Windows with MSVC.

Cross-compilation is of course also supported, but
has not been tested yet for this build.

Try with:

 $ meson --prefix=/tmp/prefix builddir
 $ ninja -C builddir
 $ ninja -C builddir test
 $ ninja -C builddir install

Build time comparison:

 - autotools: 18 seconds
 - meson: 4 seconds

http://mesonbuild.com for more information.