test-runner.py: Fix running on Visual Studio builds
On Windows, since we need to be able to find the fribidi DLL when
running the tests that make use of fribidi.exe (which is in
$(builddir)\bin), we need to prepend $(builddir)\lib into %PATH% so that
we can load the fribidi DLL that we just built when running fribidi.exe.

Also, as Visual Studio builds are run in standard Windows cmd.exe consoles,
we need to take account of the output line endings, which are '\r\n',
versus '\n' on *NIX builds.  Since the reference files are in various
encodings (hence we need to deal with binary comparison), we can't just use
universal_newlines when doing subprocess.check_output(), as that will
put us into text mode.

Finally, fix a typo in the error message when the output and reference
don't match.