remove useless library arguments applied to executables
Setting the library symbol visibility when compiling executables
accomplishes nothing. This is only filled out as containing a value on
Linux / GNU-like compilers, and in this case the fallback value of
absolutely nothing is entirely sufficient. It simply does not matter --
we aren't building a library.

All this did was make command lines longer and a bit less readable.