simplify visibility handling
The previous handling of visibility was overengineered. It suffices to
always set up visibility on GCC, which means this can be dropped from
meson.build while simultaneously being shorter in the header setup.

Meson natively knows when to pass -fvisibility=hidden, without any
conditionals; all that is needed is to ensure that the visibility
annotations for public symbols are always set.

This also avoids some inefficiencies or bugs, such as:
- passing -fvisibity for all targets, not just the ones that actually
  use it
- misdetecting whether the compiler supports -fvisibility=hidden, by
  checking whether the command line argument is accepted rather than
  whether `cc.has_function_attribute()` can compile code which makes use
  of the attribute
- passing -DFRIBIDI_BUILD on non-Windows platforms, where it isn't
  actually used