meson: Override dependencies to improve usage as a subproject
With this change, fribidi can be consumed as a subproject without
making any changes to the build files of a project. All you need to do
is provide a wrap file with a `[provide]` section:

https://mesonbuild.com/Wrap-dependency-system-manual.html#provide-section

This is also necessary because otherwise projects need to hard-code
the subproject name, which might be `fribidi` when using `wrap-git` or
`fribidi-1.0.12` when using `wrap-file` (to build from a release
tarball). This can cause conflicts between different subprojects that
consume fribidi differently.

Other projects like glib, cairo, pango, etc already do this.

Bumped the minimum version of Meson to 0.54 because that's the release
that added this feature. It was released almost three years ago, so
should be fine.